backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
pillow==10.1.0
aiofiles==23.2.1
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
